Updated to zsv 1.3.0 (#62)

+++ csv.mod/zsv/scripts/ci-build.sh
@@ -19,6 +19,10 @@ if [ "$RUN_TESTS" != true ]; then
   RUN_TESTS=false
 fi
 
+if [ "$SKIP_BUILD" != true ]; then
+  SKIP_BUILD=false
+fi
+
 if [ "$SKIP_ZIP_ARCHIVE" != true ]; then
   SKIP_ZIP_ARCHIVE=false
 fi
@@ -26,6 +30,8 @@ fi
 if [ "$SKIP_TAR_ARCHIVE" != true ]; then
   SKIP_TAR_ARCHIVE=false
 fi
+
+WITHOUT_SIMD=${WITHOUT_SIMD:-false}
 
 #JQ_DIR="$PWD/jq"
 #JQ_PREFIX="$JQ_DIR/build"
@@ -37,10 +43,14 @@ echo "[INF] Building and generating artifacts"
 echo "[INF] PWD:              $PWD"
 echo "[INF] PREFIX:           $PREFIX"
 echo "[INF] CC:               $CC"
+echo "[INF] CFLAGS:           $CFLAGS"
 echo "[INF] LDFLAGS:          $LDFLAGS"
 echo "[INF] MAKE:             $MAKE"
 echo "[INF] RUN_TESTS:        $RUN_TESTS"
+echo "[INF] STATIC_BUILD:     $STATIC_BUILD"
 echo "[INF] ARTIFACT_DIR:     $ARTIFACT_DIR"
+echo "[INF] WITHOUT_SIMD:     $WITHOUT_SIMD"
+echo "[INF] SKIP_BUILD:       $SKIP_BUILD"
 echo "[INF] SKIP_ZIP_ARCHIVE: $SKIP_ZIP_ARCHIVE"
 echo "[INF] SKIP_TAR_ARCHIVE: $SKIP_TAR_ARCHIVE"
 #echo "[INF] JQ_DIR:           $JQ_DIR"
@@ -55,10 +65,22 @@ echo "[INF] Listing compiler version [$CC]"
 
 echo "[INF] Configuring zsv"
 # CFLAGS="-I$JQ_INCLUDE_DIR" LDFLAGS="-L$JQ_LIB_DIR"
-./configure \
-  --prefix="$PREFIX" \
-  --disable-termcap
-#  --enable-jq
+
+if [ "$WITHOUT_SIMD" = true ]; then
+  ./configure \
+    --prefix="$PREFIX" \
+    --disable-termcap \
+    --arch=none \
+    --try-avx512=no \
+    --force-avx2=no \
+    --force-avx=no \
+    --force-sse2=no
+else
+  ./configure \
+    --prefix="$PREFIX" \
+    --disable-termcap
+  # --enable-jq
+fi
 
 if [ "$RUN_TESTS" = true ]; then
   echo "[INF] Running tests"
@@ -66,8 +88,8 @@ if [ "$RUN_TESTS" = true ]; then
   "$MAKE" test
   echo "[INF] Tests completed successfully!"
 
-  if [ "$CC" = "musl-gcc" ] && [ "$(echo "$LDFLAGS" | grep -- "-static")" != "" ]; then
-    echo "[WRN] Dynamic extensions are not supported with static musl build! Skipping tests..."
+  if [ "$(echo "$LDFLAGS" | grep -- "-static")" != "" ] || [ "$STATIC_BUILD" = "1" ]; then
+    echo "[WRN] Dynamic extensions are not supported with static builds! Skipping tests..."
   else
     echo "[INF] Configuring example extension and running example extension tests"
     echo "[INF] (cd app/ext_example && $MAKE CONFIGFILE=../../config.mk test)"
@@ -76,32 +98,34 @@ if [ "$RUN_TESTS" = true ]; then
   fi
 fi
 
-echo "[INF] Building"
-rm -rf build "$PREFIX" /usr/local/etc/zsv.ini
-"$MAKE" install
-tree -h "$PREFIX"
-echo "[INF] Built successfully!"
+if [ "$SKIP_BUILD" = false ]; then
+  echo "[INF] Building"
+  rm -rf build "$PREFIX" /usr/local/etc/zsv.ini
+  "$MAKE" install
+  tree "$PREFIX"
+  echo "[INF] Built successfully!"
 
-mkdir -p "$ARTIFACT_DIR"
+  mkdir -p "$ARTIFACT_DIR"
 
-if [ "$SKIP_ZIP_ARCHIVE" = false ]; then
-  ZIP="$PREFIX.zip"
-  echo "[INF] Compressing [$ZIP]"
-  cd "$PREFIX"
-  zip -r "$ZIP" .
-  ls -Gghl "$ZIP"
-  cd ..
-  mv "$PREFIX/$ZIP" "$ARTIFACT_DIR"
-  echo "[INF] Compressed! [$ZIP]"
-fi
+  if [ "$SKIP_ZIP_ARCHIVE" = false ]; then
+    ZIP="$PREFIX.zip"
+    echo "[INF] Compressing [$ZIP]"
+    cd "$PREFIX"
+    zip -r "$ZIP" .
+    ls -hl "$ZIP"
+    cd ..
+    mv "$PREFIX/$ZIP" "$ARTIFACT_DIR"
+    echo "[INF] Compressed! [$ZIP]"
+  fi
 
-if [ "$SKIP_TAR_ARCHIVE" = false ]; then
-  TAR="$PREFIX.tar.gz"
-  echo "[INF] Compressing [$TAR]"
-  tar -czvf "$TAR" "$PREFIX"
-  ls -Gghl "$TAR"
-  mv "$TAR" "$ARTIFACT_DIR"
-  echo "[INF] Compressed! [$TAR]"
+  if [ "$SKIP_TAR_ARCHIVE" = false ]; then
+    TAR="$PREFIX.tar.gz"
+    echo "[INF] Compressing [$TAR]"
+    tar -czvf "$TAR" "$PREFIX"
+    ls -hl "$TAR"
+    mv "$TAR" "$ARTIFACT_DIR"
+    echo "[INF] Compressed! [$TAR]"
+  fi
 fi
 
 echo "[INF] --- [DONE] ---"
